# helpan-ai-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Helpan AI Rail
  version: '1.0.0'
  summary: Agent rail for the Kirimon Market Ventures platform — orchestration runtime, delegated authority, consent surface, and audit for AI agents acting on behalf of users across Kipkiren Pay, Identiti, and Todoku.
  description: |
    Helpan AI is the **fourth platform rail** in the Kirimon Market Ventures (KMV) stack, alongside
    Identiti (identity), Kipkiren Pay (payment), and Todoku (comms). It provides the agent orchestration
    runtime, briefing storage, matching engine, delegated authority issuance/validation/revocation, the
    Helpan Console consent surface, audit log, and OAuth scope catalogue for third-party agents.

    Per design law (Helpan AI Instruction Pack §3) and the Agentic AI Signal Scan integration (7 May 2026):
    - Helpan AI does NOT hold funds, send messages, or store identity material. Every primitive that
      belongs to KP, Identiti, or Todoku is invoked via the relevant rail's API.
    - Every agent action carries a scoped, time-bounded, revocable **delegated authority token**.
    - Every issued delegated authority is visible in the Helpan Console with one-tap revocation.
    - Per-call validation of delegated authorities against this rail's `POST /authorities/{id}/validate`
      endpoint is mandatory for relying parties (KP, Todoku, consuming-app servers).
    - `initiated_by` and `actor` claims propagate cross-rail per Schema Appendix amendments §A on
...

...
  contact:
...
servers:
...
security:
  - HmacSha256: []
  - BearerCustomer: []
  - BearerService: []
paths:
...
  /briefings:
    post:
      tags: [Briefings]
      summary: Create a briefing on the user's behalf
      operationId: createBriefing
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBriefingRequest'
      responses:
        '201':
          description: Briefing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingResponse'
        '400':
          $ref: '#/components/responses/RequestInvalid'
        '401':
          $ref: '#/components/responses/AuthHmacInvalid'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '429':
          $ref: '#/components/responses/RateExceeded'

    get:
      tags: [Briefings]
      summary: List the authenticated user's briefings (filtered by app and status)
      operationId: listBriefings
      parameters:
        - name: app_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, expired, revoked]
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Briefings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingListResponse'
        '401':
          $ref: '#/components/responses/AuthHmacInvalid'
  /briefings/{briefing_id}:
    parameters:
      - $ref: '#/components/parameters/BriefingId'
    get:
      tags: [Briefings]
      summary: Read a single briefing
      operationId: getBriefing
      responses:
        '200':
          description: Briefing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Briefings]
      summary: Update a briefing (mutable fields only)
      operationId: updateBriefing
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBriefingRequest'
      responses:
        '200':
          description: Briefing updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingResponse'
        '400':
          $ref: '#/components/responses/RequestInvalid'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Briefings]
      summary: Revoke a briefing (soft-delete; audit-logged)
      operationId: revokeBriefing
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      responses:
        '204':
          description: Briefing revoked
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # DELEGATED AUTHORITIES
  # ==========================================================================
...
  /events/ingest:
    post:
      tags: [Events]
      summary: Publish an event for matching against user briefings
      operationId: ingestEvent
      description: |
        Consuming apps publish domain events that may match user briefings (e.g. Lunch Drop
        publishes "PowerMama X is offering today's special"). Helpan AI matches events against
        active briefings and emits `BriefingMatched` events on the platform Kafka bus, which the
        consuming app consumes to surface to the user.
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEventRequest'
      responses:
        '202':
          description: Event accepted for matching
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEventResponse'
        '400':
          $ref: '#/components/responses/RequestInvalid'
        '429':
          $ref: '#/components/responses/RateExceeded'

  # ==========================================================================
  # OAUTH — third-party agent scope catalogue
  # ==========================================================================
  /oauth/scopes:
    get:
      tags: [OAuth]
      summary: List available OAuth scopes (third-party agent registration uses this)
      operationId: listOauthScopes
      security: []
      parameters:
        - name: rail
          in: query
          schema:
            type: string
            enum: [helpan, kipkiren_pay, identiti, todoku, lunchdrop, chapaa, klokd]
      responses:
        '200':
          description: OAuth scope catalogue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OauthScopeListResponse'

    post:
      tags: [OAuth]
      summary: Define a new OAuth scope (admin only)
      operationId: createOauthScope
      security:
        - HmacSha256: [helpan:admin]
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOauthScopeRequest'
      responses:
        '201':
          description: Scope created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OauthScopeResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==========================================================================
  # OPERATOR — admin endpoints (FIDO2/WebAuthn auth required per Reboot Pack §10)
  # ==========================================================================
...
components:
...
  securitySchemes:
    HmacSha256:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Format: `Helpan-HMAC-SHA256 app_id={app_id}, signature={base64_sig}`.
        Canonical string (identical across all four rails per App Integration Guide §5.7):
        `{METHOD}\n{PATH_AND_QUERY}\n{CONTENT_TYPE}\n{X-Helpan-Timestamp}\n{SHA256_HEX(body)}`.
        mTLS additionally required at the transport layer.

    BearerCustomer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued customer JWT. Audience must include this rail's audience
        (`https://api.helpan.co.ke`). Carries `sub` (Account UUID), `tier`, `actor`,
        `initiated_by`, `auth_factors`, `exp`. Per Identiti Rail Contract Amendment §A.4:
        30-min standard / 5-min elevated TTL.

    BearerService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued service JWT for service-to-service authentication
        (consuming-app servers calling Helpan AI on behalf of the user). 15-min standard /
        5-min elevated TTL per Identiti Rail Contract Amendment §A.4.

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
  parameters:
...
    XIdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        UUIDv4 per logical operation. Apps SHOULD use the UUIDv5-from-business-op-id pattern
        per App Integration Guide §7.6. 24-hour idempotency window.
...
    Traceparent:
      name: Traceparent
      in: header
      required: false
      schema: { type: string, pattern: '^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$' }
      description: W3C Trace Context. Propagate unchanged across rails for cross-rail correlation.
...
    BriefingId:
      name: briefing_id
      in: path
      required: true
      schema: { type: string, pattern: '^brf_[0-9A-HJKMNP-TV-Z]{26}$' }
      description: Briefing identifier (`brf_<ULID>`).
...
  schemas:
...
    Meta:
      type: object
      required: [request_id, timestamp, schema_version]
      properties:
        request_id:
          type: string
          description: ULID
        timestamp:
          type: string
          format: date-time
        schema_version:
          type: string
          enum: ['1.0']
    SuccessEnvelope:
      type: object
      required: [ok, data, meta]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          description: Operation-specific payload
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required: [ok, error, meta]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code (e.g. `AUTH_HMAC_INVALID`)
        message:
          type: string
          description: Human-readable; not stable across versions
        field:
          type: string
          nullable: true
        detail:
          type: object
          additionalProperties: true
        documentation_url:
          type: string
          format: uri

    # ---- shared atoms ------------------------------------------------------
...
    AccountUuid:
      type: string
      pattern: '^acc_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
...
    BriefingType:
      type: string
      enum: [alert, standing_basket, scheduled_action, threshold_watch]
    BriefingStatus:
      type: string
      enum: [active, paused, expired, revoked]
    Briefing:
      type: object
      required: [id, account_uuid, app_id, briefing_type, status, intent, created_at]
      properties:
        id:
          type: string
          pattern: '^brf_[0-9A-HJKMNP-TV-Z]{26}$'
        account_uuid:
          $ref: '#/components/schemas/AccountUuid'
        app_id:
          type: string
        agent_id:
          type: string
          pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$'
        briefing_type:
          $ref: '#/components/schemas/BriefingType'
        status:
          $ref: '#/components/schemas/BriefingStatus'
        intent:
          type: object
          additionalProperties: true
          description: Structured intent — schema varies by app and briefing_type
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateBriefingRequest:
      type: object
      required: [account_uuid, app_id, briefing_type, intent]
      properties:
        account_uuid: { $ref: '#/components/schemas/AccountUuid' }
        app_id: { type: string }
        agent_id:
          type: string
          pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$'
        briefing_type: { $ref: '#/components/schemas/BriefingType' }
        intent:
          type: object
          additionalProperties: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    UpdateBriefingRequest:
      type: object
      properties:
        status: { $ref: '#/components/schemas/BriefingStatus' }
        intent:
          type: object
          additionalProperties: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    BriefingResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Briefing'
    BriefingListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Briefing'
                next_cursor:
                  type: string
                  nullable: true

    # ---- delegated authorities ---------------------------------------------
...
    IngestEventRequest:
      type: object
      required: [event_type, app_id, payload]
      properties:
        event_type:
          type: string
          description: 'App-defined event type (e.g. `lunchdrop.zone_feed.broadcast`)'
        app_id:
          type: string
        account_uuid:
          $ref: '#/components/schemas/AccountUuid'
          description: Optional — if event is account-scoped
        payload:
          type: object
          additionalProperties: true
        published_at:
          type: string
          format: date-time
        app_correlation_id:
          type: string
    IngestEventResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                event_id:
                  type: string
                  pattern: '^evt_[0-9A-HJKMNP-TV-Z]{26}$'
                accepted_at:
                  type: string
                  format: date-time

    # ---- OAuth scopes -----------------------------------------------------
    OauthScope:
      type: object
      required: [id, name, description, rail, default_grantable]
      properties:
        id:
          type: string
          description: 'e.g. `kipkiren.read.balance`, `chapaa.mmf.rebalance`'
        name:
          type: string
        description:
          type: string
        rail:
          type: string
          enum: [helpan, kipkiren_pay, identiti, todoku, lunchdrop, chapaa, klokd]
        default_grantable:
          type: boolean
          description: |
            False for behavioural-detail scopes per scan §6.3 (behavioural-data containment).
            False scopes require additional friction in the Helpan Console grant flow.
        elevation_friction:
          type: string
          enum: [low, medium, high]
        category:
          type: string
          enum: [read_aggregate, read_behavioural, write_money, write_comms, write_identity, admin]
    OauthScopeListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/OauthScope' }
    OauthScopeResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/OauthScope' }
    CreateOauthScopeRequest:
      type: object
      required: [id, name, description, rail, default_grantable, category]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        rail: { type: string }
        default_grantable: { type: boolean }
        elevation_friction: { type: string, enum: [low, medium, high] }
        category: { type: string }

    # ---- agents -----------------------------------------------------------
...
  responses:
    AuthHmacInvalid:
      description: HMAC signature invalid, missing, or timestamp out of tolerance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authentication succeeded but scope or tenant status forbids the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RequestInvalid:
      description: Request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    IdempotencyConflict:
      description: Idempotency key reused with different request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateExceeded:
      description: Per-app rate or velocity envelope exceeded
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

# ============================================================================
# WEBHOOKS
# ============================================================================
...